kernel: aarch32: split arch.c into different files (in arch/)
