+++ src/kernel/arch/aarch32/Makefile.include
@@ -28,6 +28,7 @@ QEMU_OPTIONS = -M versatilepb -serial stdio
 
 libk_c_files += $(wildcard $(kernel_src_dir)/core/arm/*.c)
 libk_c_files += $(wildcard $(arch_src)/*.c)
+libk_c_files += $(wildcard $(arch_src)/arch/*.c)
 libk_c_files += $(wildcard $(arch_src)/core/*.c)
 libk_c_files += $(wildcard $(arch_src)/mmu/*.c)
