+++ src/kernel/arch/aarch32/arch/kshell.c
@@ -1,0 +1,6 @@
+#include <arch/kshell.h>
+
+void arch_selftest()
+{
+  // TODO:
+}
