+++ src/kernel/arch/aarch32/arch/panic.c
@@ -1,0 +1,16 @@
+#include <arch/panic.h>
+
+#include <arvern/utils.h>
+
+void arch_kernel_load_symbols(uint64_t addr, uint64_t size)
+{
+  UNUSED(addr);
+  UNUSED(size);
+
+  // TODO:
+}
+
+void arch_kernel_dump_stacktrace()
+{
+  // TODO:
+}
